requirements.txt
pip
requests
PyYAML
kafka-python
psycopg2-binary